Criação do DTO de Cursos, tratamento de exceptions, criação de enums e mappers para categoria e refatoração no serviço e no controller

+++ crud-spring/src/main/java/com/nathan/crud_spring/exceptions/RecordNotFoundException.java
@@ -3,7 +3,7 @@ package com.nathan.crud_spring.exceptions;
 public class RecordNotFoundException extends RuntimeException {
     private static final long serialVersionUID = 1L;
 
-    public RecordNotFoundException(Long id) {
+    public RecordNotFoundException(Integer id) {
         super("Registro não encontrado com o id: " + id);
     }
