+++ crud-spring/src/main/java/com/nathan/crud_spring/repository/CourseRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.nathan.crud_spring.model.Course;
 
 @Repository
-public interface CourseRepository extends JpaRepository <Course, Long> {
-    
+public interface CourseRepository extends JpaRepository <Course, Integer> {
+	
 }
